Lots of refactors in the codebase to make way for a new ModelBaseline to work
